+++ src/animation/utils/transitions.ts
@@ -38,18 +38,13 @@ export function isTransitionDefined({
  * Convert Framer Motion's Transition type into Popmotion-compatible options.
  */
 export function convertTransitionToAnimationOptions<T>({
-    yoyo,
-    loop,
-    flip,
     ease,
     times,
     ...transition
 }: PermissiveTransitionDefinition): AnimationOptions<T> {
     const options: AnimationOptions<T> = { ...transition }
 
-    if (times) {
-        ;(options as any).offset = times
-    }
+    if (times) options["offset"] = times
 
     /**
      * Convert any existing durations from seconds to milliseconds
@@ -72,14 +67,6 @@ export function convertTransitionToAnimationOptions<T>({
      * Support legacy transition API
      */
     if (transition.type === "tween") options.type = "keyframes"
-    if (yoyo) {
-        options.repeatType = "reverse"
-    } else if (loop) {
-        options.repeatType = "loop"
-    } else if (flip) {
-        options.repeatType = "mirror"
-    }
-    options.repeat = loop || yoyo || flip || transition.repeat
 
     /**
      * TODO: Popmotion 9 has the ability to automatically detect whether to use